Updated for my changes

+++ src/main/java/com/authentication/implementation/repository/CartProductRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CartProductRepository extends JpaRepository<CartProduct, Long> {
+public interface CartProductRepository extends JpaRepository<CartProduct, Integer> {
 }
